Put spam protection in game cog as a listener (#24)
* Remove spam protection from main file

* Moved spam protection to game cog file
